// src/app/models/container-with-iframe.ts
export class ContainerWithIframeInfo {
    header: string = '';
    iframe_upper_img_src: string = '';
    iframe_src: string = '';
    iframe_title: string = '';
    content: string = '';
    borderColorCode: string = '';
    headerColorCode: string = '';
    show_add_my_routine_btn: boolean = false;

    constructor() {}
}